pyproject.toml
[project]
name = "text-to-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.0",
    "faiss-cpu>=1.13.2",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
